update submodules; minor config changes
- alacritty font size reduced; use higher dpi instead

+++ bin/bin/audio-rename.sh
@@ -6,11 +6,23 @@ if [ "$1" = "--dry" ]; then
     shift
 fi
 
+GOOD='a-zA-z0-9 _\-.,:'
 for f in "$@"; do
-    name=$(mediainfo --Inform="General;%Track/Position%-%Title%.mp3" "$f")
+    ext="${f##*.}"
+    name=$(mediainfo --Inform="General;%Track/Position%-%Title%.$ext" "$f" | \
+               tr 'è' 'e' | \
+               tr 'È' 'E' | \
+               tr 'à' 'a' | \
+               tr 'À' 'A' | \
+               tr -d -c "$GOOD" | tr ':' ',')
     if [ $dry -eq 1 ]; then
         echo "$f --> $name"
     else
-        mv "$f" "$name"
+        if [ -e "$name" ]; then
+            echo "Already correct name: $name"
+        else
+            echo "Move $f --> $name"
+            mv "$f" "$name"
+        fi
     fi
 done
